Backend FInished(Almost)
백엔드 파트 완료 (추가 수정 부분만 남아 있음)

+++ src/controllers/videoController.js
@@ -1,10 +1,11 @@
 import videoConstructor from "../models/Video";
+import userConstructor from "../models/User";
 
-export const globalHotVideo=async(req,res)=>{
+export const rootHotVideo=async(req,res)=>{
     const videoConstructors=await videoConstructor.find({}).sort({ creationAt: "asc"});
     return res.render(`home`, { pageTitle: `Home`, videoConstructors}); 
 }
-export const globalSearchVideo=async(req,res)=>{
+export const rootSearchVideo=async(req,res)=>{
     const { keyword }=req.query;
     let videoSearch=[];
     if(keyword){
@@ -20,22 +21,37 @@ export const globalSearchVideo=async(req,res)=>{
 
 export const seeVideo=async(req,res)=>{
     const { id } = req.params;
-    const videoNow=await videoConstructor.findById(id);
-    // console.log(videoNow);
+    const videoNow=await videoConstructor.findById(id).populate(`owner`);
     if (!videoNow){
-        return res.render(`404`, { pageTitle: `Video not found.`, videoNow});
+        return res.render(`404`, {
+            pageTitle: `Video not found.`,
+            videoNow
+        });
     } else {
-        return res.render(`seevideo`, { pageTitle: `See: ${videoNow.title}`, videoNow});
+        return res.render(`seevideo`, {
+            pageTitle: `See: ${videoNow.title}`,
+            videoNow,
+        });
     }
 }
 // Edit 편집
 export const getEditVideo=async(req,res)=>{
-    const { id } = req.params;
+    const {
+        params: { id },
+        session: { user: { _id }},
+    }=req;
+
     const videoNow=await videoConstructor.findById(id);
     if (!videoNow) {
-        return res.render(`404`, { pageTitle: `Video not found.`, videoNow});
+        return res.status(404).render(`404`, {
+            pageTitle: `Video not found.`, videoNow
+        });
+    } else if (String(videoNow.owner) !== _id) {
+        return res.status(403).redirect("/");
     } else {
-        return res.render(`geteditvideo`, { pageTitle: `Editing: ${videoNow.title}`, videoNow});
+        return res.render(`geteditvideo`, {
+            pageTitle: `Editing: ${videoNow.title}`, videoNow
+        });
     }
 }
 export const postEditVideo=async(req,res)=>{
@@ -44,7 +60,7 @@ export const postEditVideo=async(req,res)=>{
     const videoNow=await videoConstructor.exists({_id:id});
     // video object 를 받는 것이 아니라 true or false 만 받는 것.
     if (!videoNow) {
-        return res.render(`404`, { pageTitle: `Video not found.`, videoNow});
+        return res.status(404).render(`404`, { pageTitle: `Video not found.`, videoNow});
     } else {
         await videoConstructor.findByIdAndUpdate(id, {
             title,
@@ -56,8 +72,25 @@ export const postEditVideo=async(req,res)=>{
 }
 // Delete 삭제
 export const deleteVideo=async(req,res)=>{
-    const { id }=req.params;
-    await videoConstructor.findByIdAndDelete(id);
+    const {
+        params: { id },
+        session: { user: { _id }},
+    }=req;
+
+    // const videoNow=await videoConstructor.findByIdAndDelete(id);
+
+    const videoNow=await videoConstructor.findById(id);
+    if (!videoNow) {
+        return res.status(404).render(`404`, {
+            pageTitle: `Video not found.`, videoNow
+        });
+    } else if (String(videoNow.owner) !== _id) {
+        return res.status(403).redirect(`/`);
+    } else {
+        await videoConstructor.findByIdAndDelete(id);
+        return res.redirect(`/`);
+    }
+
     return res.redirect("/");
 }
 // Uplaod 업로드
@@ -65,18 +98,26 @@ export const getUploadVideo=(req,res)=>{
     return res.render(`getuploadvideo`, { pageTitle: 'Upload Video'});
 }
 export const postUploadVideo=async(req,res)=>{
-    const { title, description, hashtags }=req.body;
+    const {
+        file: { path: fileUrl },
+        session: { user: { _id } },
+        body: { title, description, hashtags },
+    }=req;
     try {
-        await videoConstructor.create({
+        const uploadVideo=await videoConstructor.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags:videoConstructor.formatHashtags(hashtags),
         });
-        // Here We will add a video to the videos array
+        const userDB = await userConstructor.findById(_id);
+        userDB.childVideo.push(uploadVideo._id);
+        userDB.save();
         return res.redirect(`/`);
     } catch (error) {
         // console.log(`videoController.js : ${error}`);
-        res.render(`getuploadvideo`, {
+        res.status(400).render(`getuploadvideo`, {
             pageTitle:'Upload Video',
             errorMessage:error,
         });
